Fixed bug with peashooter and increased cherry bomb range.

+++ PvZ/Plant.pde
@@ -27,7 +27,7 @@ public class Plant{
     return Cost;
   }
   
-  double getRefresh(){
+  int getRefresh(){
    return refresh; 
   }
